refactor: split KaitenClient into domain-specific extension files
Extract all API methods from the monolithic KaitenClient.swift into
separate extension files organized by domain (Cards, Boards, Spaces,
Columns, Lanes, Members, Tags, Comments, Checklists, Blockers,
ExternalLinks, CustomProperties, Sprints, Users). Move CardFilter
to its own top-level file with a typealias for backwards compatibility.
Core file retains only struct definition, init, and internal helpers.

Closes #347
